Separate int16_ref into separate file
